# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.LoginRequest:
    properties:
      client_id:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.LoginResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
...
  models.PasswordResetConfirmRequest:
    properties:
      otp:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - otp
    - password
    type: object
  models.PasswordResetRequest:
    properties:
      client_id:
        type: string
      username:
        type: string
    required:
    - client_id
    - username
    type: object
...
  models.SelerType:
    enum:
    - seler
    - promoter
    type: string
    x-enum-varnames:
    - SelerTypeSeler
    - SelerTypePromoter
  models.SubscriptionStatus:
    enum:
    - active
    - inactive
    - expired
    - onhold
    - paused
    - canceled
    type: string
    x-enum-varnames:
    - SubscriptionActive
    - SubscriptionInactive
    - SubscriptionExpired
    - SubscriptionOnHold
    - SubscriptionPaused
    - SubscriptionCanceled
  models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      client_id:
        type: string
      commission_percentage:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      email:
        type: string
      first_name:
        type: string
      has_active_subscription:
        type: boolean
      id:
        type: string
      institution_id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      national_id:
        type: string
      notes:
        type: string
      otp_required:
        type: boolean
      passport_number:
        type: string
      phone:
        type: string
      profile_picture:
        type: string
      referral_code:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      seler_type:
        $ref: '#/definitions/models.SelerType'
      slug:
        type: string
      specialization:
        type: string
      status:
        $ref: '#/definitions/models.UserStatus'
      subscription_status:
        $ref: '#/definitions/models.SubscriptionStatus'
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.UserRole:
    enum:
    - user
    - admin
    - super_admin
    - trainer
    - instructor
    - frontdesk
    - finance
    - seler
    - member
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
    - RoleSuperAdmin
    - RoleTrainer
    - RoleInstructor
    - RoleFrontdesk
    - RoleFinance
    - RoleSeler
    - RoleMember
  models.UserStatus:
    enum:
    - active
    - inactive
    - suspended
    - deleted
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusInactive
    - StatusSuspended
    - StatusDeleted
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive user management system with authentication and authorization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: User Management API
  version: "1.0"
paths:
...
  /users/check:
    get:
      description: Verify if user is authenticated
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check Authentication
      tags:
      - Authentication
  /users/confirm-password-reset-otp:
    post:
      consumes:
      - application/json
      description: Confirm password reset with OTP
      parameters:
      - description: Password reset confirmation
        in: body
        name: passwordResetConfirmRequest
        required: true
        schema:
          $ref: '#/definitions/models.PasswordResetConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Confirm Password Reset
      tags:
      - Authentication
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username/email and password
      parameters:
      - description: Login credentials
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: User Login
      tags:
      - Authentication
  /users/logout:
    post:
      description: Logout user and blacklist token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: User Logout
      tags:
      - Authentication
  /users/password-reset:
    post:
      consumes:
      - application/json
      description: Request password reset OTP
      parameters:
      - description: Password reset request
        in: body
        name: passwordResetRequest
        required: true
        schema:
          $ref: '#/definitions/models.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Request Password Reset
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"